Migrate RequestMatcherConfigurerTests groovy->java
Issue: gh-4939
